pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taskr"
version = "0.2.1"
description = "A CLI wrapper for TaskWarrior with enhanced UX"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Joel Yisrael", email = "schizoprada@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.7.0",
    "questionary>=1.10.0",
    "rich>=12.6.0",
    "pyyaml>=6.0",
    "pydantic>=1.10.0",
    "colorama",
    "python-dateutil"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Documentation = "https://github.com/schizoprada/taskr#readme"
Source = "https://github.com/schizoprada/taskr"
Issues = "https://github.com/schizoprada/taskr/issues"

[project.scripts]
taskr = "taskr.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/taskr"]

[tool.isort]
profile = "black"
known_first_party = ["taskr"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
